【RestAP sqlite化】マージしたらコンパイルエラーになったので修正（もう一か所あった・・・）

+++ app/src/main/java/com/v3/basis/blas/ui/fixture/fixture_view/FixtureListViewModel.kt
@@ -12,7 +12,8 @@ import java.util.concurrent.TimeUnit
 
 class FixtureListViewModel: ViewModel() {
 
-    private val disposableMap: MutableMap<Int, CompositeDisposable> = mutableMapOf()
+    //private val disposableMap: MutableMap<Int, CompositeDisposable> = mutableMapOf()
+    private val disposableMap: MutableMap<Long, CompositeDisposable> = mutableMapOf()
 
     fun clickSyncToServer(model: FixtureCellModel) {
